refactor(api): hoist the scope grant algebra to convex/lib/scopes.ts
hasScope/parseScopes/isScopeGranted move to an isomorphic, import-free module so
the Convex-side agent guard and the Node-side key path share ONE implementation
(R-3.1, same cross-boundary pattern as permissionsCore). The Convex copy is the
load-bearing one: it runs inside the mutation transaction, so a leaked agent
token calling Convex directly cannot dodge it. src/lib/api-key.ts re-exports so
existing importers don't move.

Adds the self:read / self:write vocabulary for the personal-scope surfaces, which
have no natural Resource to intersect against.
